+++ LiteXLoader/LiteXLoader/API/PlayerAPI.h
@@ -36,7 +36,7 @@ public:
     Local<Value> getInWater();
     Local<Value> getSneaking();
     Local<Value> getSpeed();
-    Local<Value> getDirection();
+    Local<Value> getRotation();
     Local<Value> getUniqueID();
 
     Local<Value> isOP(const Arguments& args);